Add git version info to version string
When building from git, it can be useful to know which version
a user has at commit level resolution. When building in a git
repository, use git describe to generate a version string. If
HEAD matches a tag, use that. Examples:

    $ ./svtplay-dl --version
    0.20.2015.10.08-3-gbd75a67

    $ git tag -a 0.20.2015.10.18
    $ make
    ...
    $ ./svtplay-dl --version
    0.20.2015.10.18

make release is also adjusted, so that it overrides the version
value when building, so that official releases still only has the
tag.