tests/hls: Specify static m3u in module instead of fetching with http
This avoids dependency on networking to run the test suite, as well as
making the test suite more robust in case of changes at the remote
location. Because of possible copyright issues, I opted not the use the
m3u file, but create a new one with similar properties.

This change also avoids hash ordering issues, by comparing the python
data structures directly, instead of a dumped json representation. There
are no guarantees that keys are ordered the same way when run in
different environments (and possibly in the same environment at
different times).