postprocess: add support to merge & remux to matroska (MKV)
This commit adds support for postprocessing to the matroska format (MKV) using either mkvmerge or ffmpeg/avconv. This is activated with the new --output-format option or by specifin output_format: mkv in the config file.

Code is still a work in progress and need to be cleaned up a bit, there might be some bugs left. 

It also change the way svtplay-dl detects if a file already exists on disk.
If name of service and video_id is included in filename it will scan for that (just as before),
if not it will just compare the filename without extension (excluding subtitles) (this might need some improvement)