Fix appveyor badge and update Solus wording
The appveyor badge was first and formost using the wrong link which meant that clicking on it sent you to look at the image and not the actual project page for it. The old image was blurry and looked bad, so I switched to the SVG badge instead. While at it I also updated the link to the new Solus site and updated the wording.